requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
absl-py==2.2.2
ADEval==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cmake==4.0.0
cycler==0.12.1
filelock==3.16.1
fsspec==2025.3.0
google-auth==2.40.2
google-auth-oauthlib==1.0.0
grpcio==1.70.0
huggingface-hub==0.30.2
idna==3.10
imagecodecs==2023.3.16
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.5.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.11.9
lit==18.1.8
logger==1.4
loguru==0.7.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.2.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.0
oauthlib==3.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.6.0.66
packaging==24.2
pandas==1.3.5
Pillow==10.4.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.7
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.12
torch==2.0.0+cu118
torchmetrics==1.5.2
torchvision==0.15.1+cu118
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
